@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
-import {HttpClientModule} from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {Routes, RouterModule, UrlSerializer} from '@angular/router';
 
 /* Libraries */
@@ -10,11 +10,11 @@ import { DynamicModule } from 'ng-dynamic-component';
 
 /* Material */
 import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material';
-import { MatGridListModule } from '@angular/material';
+import { MatCardModule } from '@angular/material/card';
+import { MatGridListModule } from '@angular/material/grid-list';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatToolbarModule } from '@angular/material';
-import { MatTabsModule } from '@angular/material';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatTabsModule } from '@angular/material/tabs';
 
 /* App components */
 import { AppComponent } from './components/app/app.component';
@@ -37,43 +37,40 @@ const routes: Routes = [
   {path: 'query', component: MainComponent},
   {path: 'query/:q', component: MainComponent},
   {path: 'detail/:filename', component: ImgDetailComponent},
+  {path: '', redirectTo: '/query', pathMatch: 'full' }, // Ruta por defecto
+  {path: '**', redirectTo: '/query' } // Ruta para manejar rutas no encontradas
 ];
 
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    ImgDetailComponent,
-    QueryResultComponent,
-    MainComponent,
-    ResultColUrlComponent,
-    ResultColImageComponent,
-    ResultColValComponent,
-    ResultColNullComponent
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    HttpClientModule,
-    RouterModule.forRoot(routes, {
-      useHash: false,
-      canceledNavigationResolution: 'computed',
-      paramsInheritanceStrategy: 'always',
-      urlUpdateStrategy: 'deferred'
-    }),
-    MatButtonModule,
-    MatCardModule,
-    MatGridListModule,
-    MatProgressSpinnerModule,
-    MatToolbarModule,
-    MatTabsModule,
-    DynamicModule.withComponents([ResultColUrlComponent, ResultColImageComponent, ResultColValComponent, ResultColNullComponent]),
-  ],
-  providers: [
-    HttpService,
-    { provide: UrlSerializer, useClass: CustomUrlSerializer }
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        ImgDetailComponent,
+        QueryResultComponent,
+        MainComponent,
+        ResultColUrlComponent,
+        ResultColImageComponent,
+        ResultColValComponent,
+        ResultColNullComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+            BrowserAnimationsModule,
+            FormsModule,
+            RouterModule.forRoot(routes, {
+                useHash: false,
+                canceledNavigationResolution: 'computed',
+                paramsInheritanceStrategy: 'always',
+                urlUpdateStrategy: 'deferred'
+            }),
+            MatButtonModule,
+            MatCardModule,
+            MatGridListModule,
+            MatProgressSpinnerModule,
+            MatToolbarModule,
+            MatTabsModule,
+            DynamicModule]
+        , providers: [
+        HttpService,
+        { provide: UrlSerializer, useClass: CustomUrlSerializer },
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
