@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ImgDetailComponent } from './img-detail.component';
 
@@ -6,7 +6,7 @@ describe('ImgDetailComponent', () => {
   let component: ImgDetailComponent;
   let fixture: ComponentFixture<ImgDetailComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ImgDetailComponent ]
     })
