@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { QueryResultComponent } from './query-result.component';
 
@@ -6,7 +6,7 @@ describe('QueryResultComponent', () => {
   let component: QueryResultComponent;
   let fixture: ComponentFixture<QueryResultComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ QueryResultComponent ]
     })
