@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ResultColNullComponent } from './result-col-null.component';
 
@@ -6,7 +6,7 @@ describe('ResultColNullComponent', () => {
   let component: ResultColNullComponent;
   let fixture: ComponentFixture<ResultColNullComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ResultColNullComponent ]
     })
