@@ -57,8 +57,8 @@ import 'core-js/es7/reflect';
 /***************************************************************************************************
  * Zone JS is required by default for Angular itself.
  */
-import 'zone.js/testing';  // Included with Angular CLI.
 
+import 'zone.js';
 
 
 /***************************************************************************************************
