@@ -1,27 +1,22 @@
-// This file is required by karma.conf.js and loads recursively all the .spec and framework files
-
-import 'zone.js/dist/long-stack-trace-zone';
-import 'zone.js/dist/proxy.js';
-import 'zone.js/dist/sync-test';
-import 'zone.js/dist/jasmine-patch';
-import 'zone.js/dist/async-test';
-import 'zone.js/dist/fake-async-test';
+import 'zone.js'; // Solo importa zone.js
 import { getTestBed } from '@angular/core/testing';
-import {
-  BrowserDynamicTestingModule,
-  platformBrowserDynamicTesting
-} from '@angular/platform-browser-dynamic/testing';
+import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 
-// Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
+// Descartar la declaración de `__karma__` si no es necesario
 declare const __karma__: any;
 
-// Prevent Karma from running prematurely.
-__karma__.loaded = function () {};
+// Evitar que Karma se ejecute prematuramente.
+if (typeof __karma__ !== 'undefined') {
+  __karma__.loaded = function () {};
+}
 
-// First, initialize the Angular testing environment.
+// Inicializar el entorno de pruebas de Angular.
 getTestBed().initTestEnvironment(
   BrowserDynamicTestingModule,
   platformBrowserDynamicTesting()
 );
-// Finally, start Karma to run the tests.
-__karma__.start();
+
+// Finalmente, inicia Karma para ejecutar las pruebas.
+if (typeof __karma__ !== 'undefined') {
+  __karma__.start();
+}
